Finalizing First Changes
I finally finalized the API, which is finally working properly. Listener class aswell. All listeners are working amazingly, thank you to cruise ship for parts of the code. Also learned how to register routes, and create a discord callback within the API. More to come in the next weekly update.